refactor: move  ReportOrganizationType
