+++ browser-extension/src/entrypoints/posts/Report/Exports/PdfReport.tsx
@@ -16,7 +16,6 @@ import {
 import { SocialNetworkName } from "@/shared/model/SocialNetworkName";
 import { getNumberOfHatefulAuthors } from "@/shared/utils/report-stats";
 import {
-  ReportOrganizationType,
   ReportQueryData,
 } from "../Stepper/BuildReport";
 import { NOTICE_UTILISATION_DATA } from "../Notice/noticeUtilisationData";
@@ -41,6 +40,7 @@ import {
   getPublicationGroups,
   GroupedData,
 } from "../ReportGroupingUtils";
+import { ReportOrganizationType } from "@/shared/model/ReportOrganizationType";
 
 const GRAY_500 = "#6b7280";
 const NEUTRAL_50 = "#FAFAFA";
