+++ browser-extension/src/entrypoints/posts/Report/Exports/reportCsv.ts
@@ -1,28 +1,18 @@
-import { getSocialNetworkName } from "@/shared/utils/post-util";
 import { Post } from "@/shared/model/post/Post";
 import { ReportQueryData } from "../Stepper/BuildReport";
 import {
-  booleanToFrenchText,
   formatDateTimeForCsv,
-  getPublicationDateRawRange,
-  publicationDateSourceText,
-  publicationDateToCsvText,
-  publicationDateTypeToText,
   reportOrganizationTypeToText,
 } from "./reportExportShared";
-import { POST_DETAIL_CSV_COLUMNS } from "@/shared/utils/csv-util";
+import {
+  buildPostCommentRow,
+  escapeCsvCell,
+  POST_DETAIL_CSV_COLUMNS,
+} from "@/shared/utils/post-csv-util";
 
 const REPORT_CSV_COLUMNS = [
   { key: "generated_at", label: "Date de génération du rapport" },
   { key: "generated_at_raw_utc", label: "Date de génération (UTC brut)" },
-  {
-    key: "post_last_analysis_at",
-    label: "Date de dernière collecte de la publication",
-  },
-  {
-    key: "post_last_analysis_at_raw_utc",
-    label: "Date de dernière collecte (UTC brut)",
-  },
   { key: "report_organization", label: "Organisation du rapport" },
   { key: "report_organization_code", label: "Organisation du rapport (code)" },
   ...POST_DETAIL_CSV_COLUMNS,
@@ -60,71 +50,15 @@ function buildReportCsvRows(
 
   return reportQueryData.postCommentList.map((comment) => {
     const post = postsByKey.get(comment.postKey);
-    const postRawDateRange = post
-      ? getPublicationDateRawRange(post.publishedAt)
-      : null;
-    const commentRawDateRange = getPublicationDateRawRange(comment.publishedAt);
-
+    console.log("post", post);
     return {
       generated_at: generatedAt,
       generated_at_raw_utc: generatedAtRawUtc,
-      post_last_analysis_at: post
-        ? formatDateTimeForCsv(post.latestAnalysisDate)
-        : "",
-      post_last_analysis_at_raw_utc: post?.latestAnalysisDate ?? "",
       report_organization: reportOrganizationTypeToText(
         reportQueryData.reportOrganizationType,
       ),
       report_organization_code: reportQueryData.reportOrganizationType,
-      social_network: post ? getSocialNetworkName(post.socialNetwork) : "",
-      social_network_code: post?.socialNetwork ?? "",
-      post_id: post?.postId ?? "",
-      post_url: post?.url ?? "",
-      post_title: post?.title ?? "",
-      post_author: post?.author.name ?? "",
-      post_published_at: post ? publicationDateToCsvText(post.publishedAt) : "",
-      post_published_at_source_text: post
-        ? publicationDateSourceText(post.publishedAt)
-        : "",
-      post_published_at_type: post
-        ? publicationDateTypeToText(post.publishedAt.type)
-        : "",
-      post_published_at_raw_start_utc: postRawDateRange?.start ?? "",
-      post_published_at_raw_end_utc: postRawDateRange?.end ?? "",
-      comment_id: comment.id,
-      comment_author: comment.author.name,
-      comment_published_at: publicationDateToCsvText(comment.publishedAt),
-      comment_published_at_source_text: publicationDateSourceText(
-        comment.publishedAt,
-      ),
-      comment_published_at_type: publicationDateTypeToText(
-        comment.publishedAt.type,
-      ),
-      comment_published_at_raw_start_utc: commentRawDateRange.start,
-      comment_published_at_raw_end_utc: commentRawDateRange.end,
-      comment_text: comment.textContent,
-      comment_classification: (comment.classification ?? []).join(", "),
-      comment_classification_raw: JSON.stringify(comment.classification ?? []),
-      comment_classified_at: comment.classifiedAt
-        ? formatDateTimeForCsv(comment.classifiedAt)
-        : "",
-      comment_classified_at_raw_utc: comment.classifiedAt ?? "",
-      comment_screenshot_available: booleanToFrenchText(
-        Boolean(comment.screenshotData),
-      ),
-      comment_is_deleted: booleanToFrenchText(comment.isDeleted),
-      comment_is_new: booleanToFrenchText(comment.isNew),
-    };
+      ...buildPostCommentRow(post, comment),
+    } as ReportCsvRow;
   });
-}
-
-function escapeCsvCell(value: string): string {
-  const startsWithFormulaTrigger = /^[=+\-@]/.test(value);
-  const sanitizedValue = startsWithFormulaTrigger ? `'${value}` : value;
-
-  if (/[;"\n\r]/.test(sanitizedValue)) {
-    return `"${sanitizedValue.replaceAll('"', '""')}"`;
-  }
-
-  return sanitizedValue;
 }
