+++ browser-extension/src/entrypoints/posts/Report/ReportComment.tsx
@@ -2,7 +2,6 @@ import { Post } from "@/shared/model/post/Post";
 import { buildDataUrl, PNG_MIME_TYPE } from "@/shared/utils/data-url";
 import { cn } from "@/lib/utils";
 import { Scale } from "lucide-react";
-import { ReportOrganizationType } from "./Stepper/BuildReport";
 import {
   getLabelAnalysisComment,
   getLabelPublishedComment,
@@ -12,6 +11,7 @@ import {
   LABEL_URL,
 } from "./reportData";
 import { PostCommentWithId } from "@/shared/utils/post-util";
+import { ReportOrganizationType } from "@/shared/model/ReportOrganizationType";
 
 interface ReportCommentProps {
   comment: PostCommentWithId;
