+++ browser-extension/src/shared/model/ReportOrganizationType.ts
@@ -1,0 +1,5 @@
+
+export enum ReportOrganizationType {
+  BY_PUBLICATION = "BY_PUBLICATION",
+  BY_AUTHOR = "BY_AUTHOR",
+}
